F2 branch

+++ OLX_clone.Server/Services/BoostService/BoostService.cs
@@ -1,21 +1,22 @@
 ﻿using AutoMapper;
 using OLX_clone.Server.Data.Contracts;
 using OLX_clone.Server.Helpers;
+using OLX_clone.Server.Middleware.Exceptions;
 using OLX_clone.Server.Models;
 
 namespace OLX_clone.Server.Services.BoostService;
 
-public class BoostService: IBoostService
+public class BoostService : IBoostService
 {
     private readonly IUnitOfWork _unitOfWork;
     private readonly IMapper _mapper;
-    
+
     public BoostService(IUnitOfWork unitOfWork, IMapper mapper)
     {
         _unitOfWork = unitOfWork;
         _mapper = mapper;
     }
-    
+
     public async Task<ApiResponse<PostBoost>> CreatePostBoost(int postId, BoostPackage boostPackage)
     {
         var postBoostFromDb = await _unitOfWork.PostBoostRepository.GetPostBoostByPostId(postId);
@@ -23,7 +24,7 @@ public class BoostService: IBoostService
         {
             await _unitOfWork.PostBoostRepository.DeleteAsync(postBoostFromDb.Id);
         }
-        
+
         var postBoost = new PostBoost()
         {
             PostId = postId,
@@ -31,66 +32,68 @@ public class BoostService: IBoostService
             AvailableBoostsCount = boostPackage.BoostCount,
             TopExpiryDate = DateTime.Now.AddDays(boostPackage.TopDurationInDays),
         };
-        
+
         var postToUpdate = await _unitOfWork.PostRepository.GetAsync(postId);
-        if (postToUpdate != null)
+
+        postToUpdate.IsTop = true;
+
+        if (boostPackage.Type == BoostType.VIP)
         {
-            postToUpdate.IsTop = true;
-            
-            if(boostPackage.Type == BoostType.VIP)
-                postToUpdate.IsVip = true;
-            
-            await _unitOfWork.PostRepository.UpdateAsync(postToUpdate);
+            postToUpdate.IsVip = true;
+            postBoost.VipExpiryDate = DateTime.Now.AddDays(boostPackage.VipDurationInDays);
+        }
+
+        await _unitOfWork.PostRepository.UpdateAsync(postToUpdate);
+
+        var createdPostBoost = await _unitOfWork.PostBoostRepository.AddAsync(postBoost);
+
+        return new ApiResponse<PostBoost>
+        {
+            Data = createdPostBoost,
+            Success = true,
+            Message = "Post boost created successfully."
+        };
+    }
+
+    public async Task<ApiResponse<bool>> BoostPost(int postId)
+    {
+        var postBoost = await _unitOfWork.PostBoostRepository.GetPostBoostByPostId(postId);
+        if (postBoost == null)
+        {
+            throw new NotFoundException("Post boost not found.");
+        }
+
+        if (postBoost.AvailableBoostsCount <= 0)
+        {
+            await _unitOfWork.PostBoostRepository.DeleteAsync(postBoost.Id);
+            return new ApiResponse<bool> { Success = false, Message = "No boosts left for this post." };
+        }
+
+        if (postBoost.TopExpiryDate < DateTime.Now)
+        {
+            postBoost.TopExpiryDate = DateTime.Now.AddDays(postBoost.NumberOfDays);
         }
         else
         {
-            return new ApiResponse<PostBoost> { Success = false, Message = "Post not found" };
+            postBoost.TopExpiryDate = postBoost.TopExpiryDate.Value.AddDays(postBoost.NumberOfDays);
         }
-        
-        if (boostPackage.Type == BoostType.VIP)
+
+        postBoost.AvailableBoostsCount--;
+
+        await _unitOfWork.PostBoostRepository.UpdateAsync(postBoost);
+
+        var post = await _unitOfWork.PostRepository.GetAsync(postId);
+        if (post == null)
         {
-            postBoost.VipExpiryDate = DateTime.Now.AddDays(boostPackage.VipDurationInDays);
+            throw new NotFoundException("Post not found.");
         }
-        
-        var createdPostBoost = await _unitOfWork.PostBoostRepository.AddAsync(postBoost);
 
-        return new ApiResponse<PostBoost> { Data = createdPostBoost, Message = "Post boost created successfully" };
-    }
-
-    
-    public async Task<ApiResponse<bool>> BoostPost(int postId)
-    {
-        try
+        if (!post.IsTop)
         {
-            var postBoost = await _unitOfWork.PostBoostRepository.GetPostBoostByPostId(postId);
-            
-            if (postBoost.AvailableBoostsCount <= 0)
-            {
-                await _unitOfWork.PostBoostRepository.DeleteAsync(postBoost.Id);
-                return new ApiResponse<bool> { Success = false, Message = "No boosts left for this post." };
-            }
-
-            if(postBoost.TopExpiryDate < DateTime.Now)
-                postBoost.TopExpiryDate = DateTime.Now.AddDays(postBoost.NumberOfDays);
-            else postBoost.TopExpiryDate = postBoost.TopExpiryDate.Value.AddDays(postBoost.NumberOfDays);
-            
-            postBoost.AvailableBoostsCount--;
-            
-            await _unitOfWork.PostBoostRepository.UpdateAsync(postBoost);
-            
-            var post = await _unitOfWork.PostRepository.GetAsync(postId);
-            
-            if (!post.IsTop)
-            {
-                post.IsTop = true;
-                await _unitOfWork.PostRepository.UpdateAsync(post);
-            }
-
-            return new ApiResponse<bool> { Success = true, Message = "Post boosted successfully." };
+            post.IsTop = true;
+            await _unitOfWork.PostRepository.UpdateAsync(post);
         }
-        catch (Exception ex)
-        {
-            return new ApiResponse<bool> { Success = false, Message = $"Error boosting post: {ex.Message}" };
-        }
+
+        return new ApiResponse<bool> { Success = true, Message = "Post boosted successfully." };
     }
 }
